Remove support for custom source file paths
This makes the code that much longer, and I think that if you go in to the trouble of customizing core features, you can probably figure out these things for yourself. We could add a check to see if any customizations has been made, and add a warning that the results may not be accurate. However, this is not in the 80% use case so I won't personally add that now.