Remove ChecksIfConfigIsUpToDate
If the config is not up to date, one of two things happens. Either, the change doesn't impact the usage much, or something goes horribly wrong (and the application might not even boot). Either way I don't think the complexity is justified at this point, especially since the implementation is not that great. Also, the documentation is clear that the configuration files must be updated when updating Hyde.