Refactor magic methods to prioritize class properties
This means that if you want to access something that truly is in the front matter, call the getter on the matter property. My intention is for that however, to be used as a data origin for constructing dynamic data which is then used by these prioritizing page matter methods.