Create method to construct schemas to control the call order
This was cause to a weird issue where some data was not present in some places. This is of course because the data had not been constructed yet. As long as each child that has a schema adds it to the new construct method (and calls the parent) this issue should be fixed.