Drops support for .jpeg in favour of .jpg in lookup table
As duplicate enum cases is not allowed. This means it will just fall back to the fileinfo extension if it exists. In the edge of edge cases where a file uses the less common .jpeg and fileinfo is not installed this may lead to broken images in the HydeRC, but I'm fine with that for now as workarounds would add complexity that is in my opinion undue at this stage.