Rename traversal question methods
Yes, this is very "left-to-right" centric, but so is the code. I'm going for clarity here, and if/when we add support for RTL this is not the place to start. I use LTR, so that's what I'm doing. Open to discussion and collaboration, but like I said, this is not the place to start.