Remove warnings from action as they add too much complexity
These proposed warnings pretty much all only kick in if the user does something that goes against the defaults (and in some cases common sense) and that that point I think it's fair enough if the developer needs to do a tiny bit of debugging. Adding dynamic checks like tagGroup validation and similar add an unjustified amount of complexity at this point in time.