Automatically support non-existent files
This was previously (if I recall correctly) designed so that if a file did not exist, no link would be made, as that could be confusing for the user as the link would not be clickable. Now that I actually think about it though, I realize that if a file that was supposed to be created was not, not being able to click a terminal link would be the least of your problems here.