Use a more graceful method for re-enabling package auto-discovery
Laravel Zero disables auto-discovery, but we want to use it, so we'll call the grandparent's method instead of the parent's, instead of the previous behaviour where Laravel enables it, Zero disables it, and we disable it again. What this change does is letting Laravel handle it, but with our custom cache path.