Move the storage directory into the app directory
This directory only used by the internal underlying Laravel application, and there is really no reason for a user to interact with it, so there's no point in having it clog up the root structure. Instead it makes more sense to have it in the app directory that already houses files for the Laravel app.