Rename HydeKernel property and accessors outputPath to outputDirectory
In the very vast majority of cases, this is indeed a directory within the project, for example '_site' or 'build'. Only in edge cases would this be a path, if for some reason one would want to use a subdirectory, or use a path outside of the project like '/var/www/html' (which is officially discouraged).

This rename then makes it more intuitively clear what kind of value is expected. Furthermore, it matches the config setting `site.output_directory`