Move the torchlight.php config file to the Framework package internals
This file will usually not need to be edited as it's preconfigured, so it does not need to be bundled with the base project install. Having it in the framework package internals means we can still load the existing data, and users can publish the file if they need to modify it. Published config files should always take precedence.