Move the commands.php config file to the Framework package internals
This file will very rarely need to be edited by the user, so it does not need to be bundled with the base project install. Having it in the framework package internals means we can still load the existing data, and users can publish the file if they need to modify it. Published config files should always take precedence.