Run task discovery in the constructor
Fixes https://github.com/hydephp/develop/issues/1154

Apparently the build task discovery logic only takes around 0.152ms to execute (even less if there are no tasks)

I presume this to be because the glob/stat info is already cached by the autoloader.

Since this is pretty much nothing I see no problem with doing this in the constructor, I thought we would need to defer it but that complexity is not warranted. Furthermore, the singleton only gets instantiated when needed, meaning that the only time the constructor is called is when the class is actually needed (IE the build process where the other rendering will be much slower)

See https://github.com/hydephp/develop/issues/1154#issuecomment-1449879195