Breaking: Pluralize class name to DataCollections
When used as an instance it makes more sense to use DataCollection, but most of the time you interact with the class directly is as a facade, so DataCollections is more semantic.

In general, it is common to use plural nouns for classes that represent a collection of related objects. Therefore, the second option, "DataCollections," would be more appropriate for a class that represents a collection of data collection objects.