Require the Hyde namespace to be first in the Yaml file
I think we always want to have a hyde entry required for "namespaced" options, as it can quickly lead to issues if a root config (and likely vendor ones) option has the same name as a namespace. So even just an empty hyde entry should be required to enable the namespaced feature. See https://github.com/hydephp/develop/issues/1341#issuecomment-1480005480