Revert "Add upcoming docs"
This reverts commit 4acc82cfb5ff431302dcbb2b11c5331efd47c872.

GITHUB_BASE_REF is the short branch name for the target branch during a PR trigger, and it is empty otherwise.

GITHUB_REF always has a value, but the value changes based on the context. Outside of a PR, GITHUB_REF is the "full" branch name (refs/heads/master). Inside of a PR, it is the PR branch name (refs/pull/123/merge).