Change root monorepo composer name to hyde/monorepo
This composer file has become widely different from the actual hyde/hyde composer file, so it doesn't make semantic sense for this project to be branded the same way as the hyde/hyde package. This change is not breaking, as the monorepo structure is not included in the HydePHP BC promise.