Make PublicationPage::$type readonly as it can't be persisted
Makes the PublicationPage::$type property readonly as it can't be persisted like that

There is no logic to change the type through the object, and I don't think there should be as it would add a considerable amount of complexity compared to the user just moving the file. Making it readonly makes it clear that it cannot be changed as changing the property would do nothing.