Update realtime compiler dashboard to block unsafe requests
As the dashboard is not password-protected, and it can make changes to the file system, we block any requests that are not coming from the host machine. While we are clear in the documentation that the realtime compiler should only be used for local development, we still want to be extra careful in case someone forgets.