Add improved conditional to fix override priority
As part of the v2 refactors, since routes aren't discovered yet, we need to check the pages directly.

This also means that custom search pages are now not built by the build:search command. We could change it so that it is, but we don't know if the custom page is a custom search feature, or a page that documents a search feature. So we don't know if this is what the user would want/expect.

Additionally, the kernel API changes to needing to put the overrides in the pages collection instead of the route collection. This is also a very very minor implementation change with minimal impact.