Add a head stack to the head component
Adds a `@head` stack to the `head.blade.php` component. My original plan was to move the existing meta stack, but this makes more sense and is more compatible. It allows code to be pushed to just before the end of the `head` component, just like in the scripts component.