Remove unnecessary default values from config call
There is no reason to expect things like "index" to be a dropdown group. Posts maybe, if for some reason a user wants to list all posts in a dropdown but at that point they need to make other changes so I see no reason to have an extra line for something so uncommon.