Remove undocumented `sitemap.dynamic_priority` config option
Either we add something like this for the changefreq logic, or we remove it, and I think it can be removed since it's not documented, and it's such an uncommon use case. And if one is unhappy with the generation, a better way is to customize the class to add custom logic.