Remove the HydeFront version and CDN URL configuration options
I can't see a good reason to want to change the version to load assets that are unlikely to work with the versions in the views in the project Hyde version. Changing the CDN URL could be useful if you want to change the host to another CDN provider, but this current system is way too convoluted for that.