Breaking: Rename `Asset::hasMediaFile` to `Asset::exists`
Also considered `has`, but exists is generally better for clarity, as it directly indicates that you're checking for the presence of a file. has can be more ambiguous, as it might imply ownership or possession rather than existence. Therefore, exists is often preferred for methods that check for the presence of an item. It's also what the Laravel Filesystem facade uses