Don't throw if the directory does not exist
If this was primarily a code API feature I would want this, but from our code (which is why we added this method) we want to be graceful if directories do not exist. If we throw, then we need to catch or add a precheck in a bunch of other places.